add KELP_CONFIG env variable support

+++ pkg/rm/rm.go
@@ -2,17 +2,20 @@ package rm
 
 import (
 	"crhuber/kelp/pkg/config"
+	"crhuber/kelp/pkg/utils"
 	"fmt"
 	"os"
 	"path/filepath"
 )
 
 func RemoveBinary(binary string) error {
-	fmt.Printf("\nRemoving binary %s...", binary)
 	binaryPath := filepath.Join(config.KelpBin, binary)
-	err := os.Remove(binaryPath)
-	if err != nil {
-		return err
+	if utils.FileExists(binaryPath) {
+		fmt.Printf("\nRemoving binary %s...", binary)
+		err := os.Remove(binaryPath)
+		if err != nil {
+			return err
+		}
 	}
 	return nil
 }
